refactor: CityRanking 페이지 컴포넌트 리팩토링
